apps/backend/pyproject.toml
[project]
name = "Enterprise-Rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <3.14"
dependencies = [
    "torch",
    "torchvision",
    "docling>=2.77.0",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.135.1",
    "fastembed>=0.7.4",
    "google-api-python-client>=2.192.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "libsql>=0.1.11",
    "llama-index-core>=0.14.15",
    "llama-index-embeddings-fastembed>=0.5.0",
    "llama-index-llms-ollama>=0.9.1",
    "llama-index-llms-openrouter>=0.4.1",
    "llama-index-node-parser-docling>=0.4.2",
    "llama-index-readers-docling>=0.4.2",
    "llama-index-readers-file>=0.5.6",
    "llama-index-vector-stores-qdrant>=0.9.1",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.11.0",
    "pyrefly>=0.56.0",
    "python-multipart>=0.0.22",
    "qdrant-client>=1.17.0",
    "qdrant-client>=1.17.0",
    "rq>=2.0.0",
    "redis>=5.2.1",
    "pydantic-settings>=2.13.1",
    "opencv-python-headless>=4.11.0.86",
    "sentence-transformers>=5.3.0",
    "sqlalchemy>=2.0.48",
    "psycopg2-binary>=2.9.11",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
torchvision = { index = "pytorch-cpu" }

[tool.uv]
override-dependencies = [
    "opencv-python-headless>=4.11.0.86",
]
